notes on venus as a base (50 - 60 char)
The first linde is a very short description of the commit and conventionally should be around 50-60 char. Below it, and after skipping a line, there is room for
a more detailed descriotion of the commit.

MAybe Mars is not suitable for a base, maybe venus is better..

Add equations if those are used in the code to commit, put context and explanations, not the code itself. the commit history and the comments on the commits should give enough information to allow other users to understand what the purpose of the code  is and which changes have been made.
Try  avoiding using stuff like "changes" as the only comment in the commit message..